// src/redux/action-types/events/events.action-type.ts
export enum EventAcionType {
  FETCH_EVENTS = 'FETCH_EVENTS',
  FETCH_EVENTS_SUCCESS = 'FETCH_EVENTS_SUCCESS',
  FETCH_EVENTS_ERROR = 'FETCH_EVENTS_ERROR',

  CREATE_EVENT = 'CREATE_EVENT',
  CREATE_EVENT_SUCCESS = 'CREATE_EVENT_SUCCESS',
  CREATE_EVENT_ERROR = 'CREATE_EVENT_ERROR',

  FETCH_EVENT_DETAILS_BY_EVENT_ID = 'FETCH_EVENT_DETAILS_BY_EVENT_ID',
  FETCH_EVENT_DETAILS_BY_EVENT_ID_SUCCESS = 'FETCH_EVENT_DETAILS_BY_EVENT_ID_SUCCESS',
  FETCH_EVENT_DETAILS_BY_EVENT_ID_ERROR = 'FETCH_EVENT_DETAILS_BY_EVENT_ID_ERROR',
}